[field] Rename contexts directory
