+++ packages/@sanity/field/src/diff/components/FieldChange.tsx
@@ -2,7 +2,7 @@ import {useDocumentOperation} from '@sanity/react-hooks'
 import classNames from 'classnames'
 import React, {useCallback, useContext, useState} from 'react'
 import {undoChange} from '../changes/undoChange'
-import {DiffContext} from '../context/DiffContext'
+import {DiffContext} from '../contexts/DiffContext'
 import {FieldChangeNode, OperationsAPI} from '../../types'
 import {ChangeBreadcrumb} from './ChangeBreadcrumb'
 import {DiffErrorBoundary} from './DiffErrorBoundary'
